onpremweb_aws/requirements.txt
#onpremweb/requirements.txt
Django>=3.2,<4.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-dotenv
django-extensions
gunicorn>=21.2.0
Pillow
#react-toastify
django-storages
boto3
